moved functionality to remove comments from rst back to readme since it's no longer a workaround
